Split IcuVersionInfo into its own file
